+++ doc/version_history.rst
@@ -6,6 +6,12 @@
 Version History
 ###############
 
+v1.1.7
+------
+
+* Remove the workaround code of backward compatibility.
+* Overwrite the ``RotatorCsc.begin_disable()`` to do the state transition of controller.
+
 v1.1.6
 ------